Update: Move Cypress helper functions into utils file.
